+++ projects/swimlane/ngx-ui/src/lib/components/card/card.component.ts
@@ -1,4 +1,12 @@
-import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output, ViewEncapsulation } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  EventEmitter,
+  Input,
+  Output,
+  ViewEncapsulation,
+  HostBinding
+} from '@angular/core';
 
 import { CardStatus } from './card-status.enum';
 import { CardOrientation } from './card-orientation.enum';
@@ -9,17 +17,32 @@ import { CardAppearance } from './card-appearance.enum';
   selector: 'ngx-card',
   templateUrl: './card.component.html',
   styleUrls: ['./card.component.scss'],
-  host: {
-    class: 'ngx-card',
-    '[class.ngx-card-horizontal]': 'orientation === CardOrientation.Horizontal',
-    '[class.ngx-card-vertical]': 'orientation === CardOrientation.Vertical',
-    '[class.flat]': 'appearance === CardAppearance.Flat',
-    '[class.disabled]': 'disabled'
-  },
   changeDetection: ChangeDetectionStrategy.OnPush,
   encapsulation: ViewEncapsulation.None
 })
 export class CardComponent {
+  @HostBinding('class') class = 'ngx-card';
+
+  @HostBinding('class.ngx-card-horizontal')
+  get horizontal() {
+    return this.orientation === CardOrientation.Horizontal;
+  }
+
+  @HostBinding('class.ngx-card-vertical')
+  get vertical() {
+    return this.orientation === CardOrientation.Vertical;
+  }
+
+  @HostBinding('class.flat')
+  get flat() {
+    return this.appearance === CardAppearance.Flat;
+  }
+
+  @HostBinding('class.disabled')
+  get isDisabled() {
+    return this.disabled;
+  }
+
   @Input() orientation: CardOrientation = CardOrientation.Horizontal;
   @Input() disabled: boolean = false;
   @Input() status: CardStatus;
@@ -32,9 +55,6 @@ export class CardComponent {
 
   @Output() select = new EventEmitter<boolean>();
   @Output() outlineClick = new EventEmitter<void>();
-
-  readonly CardOrientation = CardOrientation;
-  readonly CardAppearance = CardAppearance;
 
   onOutlineClick($event) {
     $event.stopPropagation();